Added code to detect when a role has been deleted from a guild, if it's a countrybot role, delete all users from the role in the DB and then remove the role from the DB. Added refresh option to stats page.

+++ CountryBot/Models/RoleModel.cs
@@ -5,6 +5,6 @@ internal class RoleModel
     public int Id { get; set; }
     public ulong GuildId { get; set; }
     public ulong RoleId { get; set; }
-    public ulong CountryId { get; set; }
+    public int CountryId { get; set; }
 }
